Optimized contract
Price per Byte will be global and the same for any server
That's why it's stored in the contract

ServerCreated event is optimized, all information is already stored on the blockchain. Only id is required.

Server don't have to store the client address. Many clients to one server relation are used. It requires having another global hashmap in the future.